move response handler logic into food_responses.js
