+++ lib/ajax_requests/foods.js
@@ -1,20 +1,11 @@
 const $ = require('jquery')
+const foodResponse = require('../response_handlers/food_responses.js')
 
 $(document).ready(function(){
   $.ajax({
     type: "GET",
     url: "https://serene-sea-75169.herokuapp.com/api/v1/foods"
   })
-  .then(appendFoods)
-  .catch(errorLog)
+  .then(foodResponse.appendFoods)
+  .catch(foodResponse.errorLog)
 })
-
-const appendFoods = (posts) => {
-  posts.forEach(function(post){
-    $('table#food-table').append(`<tr><td>${post.name}</td> <td>${post.calories}</td></tr>`)
-  })
-}
-
-const errorLog = (error) => {
-  console.error(error)
-}
